WL: Fixed bug and testCode.R to show off the new integer functionality

+++ testCode.R
@@ -2,7 +2,7 @@ library(ROpenCL)
 
 a <- seq(11444777)/10
 b <- seq(11444777)
-out <- rep(0, length(a))
+out <- rep(0.0, length(a))
 
 localWorkSize = 256
 globalWorkSize = ceiling(length(a)/localWorkSize)*localWorkSize
@@ -17,7 +17,7 @@ queue <- createCommandQueue(context, deviceIDs[[1]])
 inputBuf1 <- createBuffer(context, "CL_MEM_READ_ONLY", globalWorkSize, a)
 inputBuf2 <- createBuffer(context, "CL_MEM_READ_ONLY", globalWorkSize, b)
 outputBuf1 <- createBufferFloatVector(context, "CL_MEM_WRITE_ONLY", globalWorkSize)
-kernel <- "__kernel void VectorAdd(__global const float* a, __global const float* b, __global float* c, int iNumElements)
+kernel <- "__kernel void VectorAdd(__global const float* a, __global const int* b, __global float* c, int iNumElements)
 {
     // get index into global data array
     int iGID = get_global_id(0);
